+++ test/jest/helpers/getTransformationFieldGroups.js
@@ -1,33 +1,15 @@
-import {
-  getByTestId,
-  queryAllByTestId,
-  screen,
-} from '@testing-library/react';
+import { screen } from '@testing-library/react';
 
 export const getTransformationFieldGroups = () => {
-  const allGroups = screen.getAllByTestId('transformation-field-group');
+  const marcFields = screen.getAllByTestId(/transformation-marcField/);
+  const indicators1 = screen.getAllByTestId(/transformation-indicator1/);
+  const indicators2 = screen.getAllByTestId(/transformation-indicator2/);
+  const subfields = screen.getAllByTestId(/transformation-subfield/);
 
-  return allGroups.map(group => ({
-    marcField: {
-      container: getByTestId(group, 'transformation-marcField'),
-      input: getByTestId(group, 'transformation-marcField').querySelector('input'),
-      isInvalid: getByTestId(group, 'transformation-marcField').classList.contains('isInvalid'),
-    },
-    indicator1: {
-      container: getByTestId(group, 'transformation-indicator1'),
-      input: getByTestId(group, 'transformation-indicator1').querySelector('input'),
-      isInvalid: getByTestId(group, 'transformation-indicator1').classList.contains('isInvalid'),
-    },
-    indicator2: {
-      container: getByTestId(group, 'transformation-indicator2'),
-      input: getByTestId(group, 'transformation-indicator2').querySelector('input'),
-      isInvalid: getByTestId(group, 'transformation-indicator2').classList.contains('isInvalid'),
-    },
-    subfield: {
-      container: getByTestId(group, 'transformation-subfield'),
-      input: getByTestId(group, 'transformation-subfield').querySelector('input'),
-      isInvalid: getByTestId(group, 'transformation-subfield').classList.contains('isInvalid'),
-    },
-    isInvalid: Boolean(queryAllByTestId(group, 'transformation-invalid').length),
-  }));
+  return {
+    marcFields,
+    indicators1,
+    indicators2,
+    subfields,
+  };
 };